Changes for code to now post provenance, enables the definition of a remote TDS and project_id across the whole stack.
