+++ env.sample
@@ -4,3 +4,4 @@ AWS_ACCESS_KEY_ID=
 AWS_SECRET_ACCESS_KEY=
 PROJECT_ID=
 TDS_URL=
+SIMSERVICE_TDS_URL=
